BE-Feat/#27-rest template

See merge request s07-bigdata-recom-sub2/S07P22A503!41

+++ server/src/main/java/com/singsong/api/controller/MemberController.java
@@ -106,7 +106,7 @@ public class MemberController {
     }
 
     @PatchMapping("/tag/add")
-    public ResponseEntity<?> memberTagAdd(@RequestParam(value = "tag") int tagId, @ApiIgnore Authentication authentication) {
+    public ResponseEntity<?> memberTagAdd(@RequestParam(value = "tag") Long tagId, @ApiIgnore Authentication authentication) {
         Member member = jwtAuthenticationUtil.jwtTokenAuth(authentication);
         tagService.addMemberTag(member, tagId);
 
@@ -114,7 +114,7 @@ public class MemberController {
     }
 
     @PatchMapping("/tag/delete")
-    public ResponseEntity<?> memberTagDelete(@RequestParam(value = "tag") int tagId, @ApiIgnore Authentication authentication) {
+    public ResponseEntity<?> memberTagDelete(@RequestParam(value = "tag") Long tagId, @ApiIgnore Authentication authentication) {
         Member member = jwtAuthenticationUtil.jwtTokenAuth(authentication);
         tagService.removeMemberTag(member, tagId);
